fix(D-16): #11-wasm-vm — Copilot R5 1 IMP (ArrayBuffer slice >4GiB)
R5 IMP (array_buffer.rs:634): `native_array_buffer_slice` cast
usize-typed `start` / `final_len` to u32 when calling
`array_buffer_view_bytes`, silently truncating slices on
ArrayBuffers > 4 GiB.  Reachable: the ArrayBuffer constructor's
`to_index_for_array_buffer` allows up to `min(2^53 - 1, usize::MAX)`,
so `new ArrayBuffer(4_300_000_000).slice(...)` triggers the bug —
returns wrong-position bytes instead of the spec-mandated range.

Structural fix per `feedback_one-issue-one-way` (single SoT):

- `array_buffer_view_bytes` signature changed from `(byte_offset:
  u32, byte_length: u32)` to `(_: usize, _: usize)`, matching
  ArrayBuffer's native indexing.  Internal wasm-backed branch uses
  `u32::try_from` guards (safe for v1 wasm: per-memory
  byte_size ≤ 2^32 — `try_from` collapses the
  inherently-unreachable overflow into the OOB empty-read
  contract).  body_data branch indexes with `usize` directly, no
  `as usize` round-trip.

- 5 call sites updated:
  - `native_array_buffer_slice`: drop the `as u32` cast (was the bug
    site).
  - `text_encoding.rs` / `blob.rs` / `image_data.rs` /
    `body_init.rs`: pass `byte_offset as usize`, `byte_length as
    usize` (zero-cost u32→usize widening on 64-bit hosts).
    `ObjectKind::TypedArray.byte_offset/byte_length` remain `u32`
    (TypedArray / DataView views are inherently u32-bounded per
    spec §23.2).

Regression test (`tests_array_buffer.rs`
`view_bytes_returns_empty_on_offset_above_u32_max`): direct call
with `byte_offset = u32::MAX as usize + 1` over a real 8-byte
ArrayBuffer — pre-fix wraps to 0 and returns 4 bytes; post-fix
returns empty (out-of-range usize index).  Avoids CI-prohibitive
multi-GiB allocation.

5261/5261 elidex-js lib tests pass.

+++ crates/script/elidex-js/src/vm/host/blob.rs
@@ -442,8 +442,8 @@ fn append_blob_part_bytes(
                     let bytes = super::array_buffer::array_buffer_view_bytes(
                         ctx.vm,
                         buffer_id,
-                        byte_offset,
-                        byte_length,
+                        byte_offset as usize,
+                        byte_length as usize,
                     );
                     out.extend_from_slice(&bytes);
                 } else {
